Remove old start-war-only-after-gaudy-pirates
Before gaudy keys were removed, this check guaranteed we wouldn't screw
ourselves out of being able to progress the L11 quest by starting the
war and shutting off the pirate zone. But that's not a worry anymore.

There are still some references to sl_gaudy and sl_swordfish but I'm not
really sure what they do.