[beta] Always adjust equipment for +/-combat
Sometimes (using maximizer-based equipment) we were running -com in
zones that we wanted to maximize +combat in, because the maximizer
really liked having the latte equipped. With this change, we run +10%
combat.

I think the reason we historically didn't do this is that some zones
required certain equipment, and we couldn't be sure that mucking with
equipped items wouldn't ruin the reason we were adventuring in the first
place. Fortunately, the new (still guarded behind beta) slEquip
functionality guarantees that we don't accidentally unequip something
important when adventuring in a zone.