acquireTotem() function in auto_util.ash to be used a wrapper for buffs and solve a mafia issue
buffMaintain (auto_util.ash:4961) would call on mafia's built in function use_skill(casts, source) for casting turtle tamer buffs without a wrapper to ensure proper handling of totem acquisition.
Under certain circumstances, such as for example not have a turtle totem and being out of meat so you can't fish for them in the sewer, this will result in autoscend erroring out
Added a function acquireTotem() and integrated that function into buffMaintain such that it will attempt to acquire a totem and if unable to will not attempt to call on mafia's use_skill(casts, source).
Tested in way of the surprising fist run, where it repeatedly errored out until I made the above changes.