Ignore duplicate accessory equips
Somewhere in our war-equipping code there's a duplicate attempt to equip
a bejeweled pledge pin.

With the newly fixed "fallback to equipping to acc2 or acc1 if acc3 is full",
this means that

1. we want to equip bejeweled pledge pin, we plan to put it in acc3

2. we want to equip bejeweled pledge pin, we plan to put it in acc2

3. we want to equip the doctor bag, we plan to put it in acc1

4. we want to equip fancy boots, we have no space, print a warning, go to
   adventure

5. we crash because we're in plumbercore with no plumber weapon equipped

This commit fixes this bug by modifying step (2) to notice we've already
planned to equip the pledge pin and have that call to autoEquip()
immediately return true.