autoForbidFamiliarChange
=========

Previously:

Egg benedict method was removed as it caused some issue with both is100familiarRun() and is100familiarRun(thisOne)
is100familiarRun() only explicitly checked for egg benedict to test if you have mosquito too, and if both then it "caught" this error and "corrected" it by marking it as allowing familiar change. Only if you had egg benedict set and did not have mosquito did it correctly block familiar changing, in that it simply reached the end of the function and returned true (indicating 100familiar is set to something).

the specific familiar version inherited this problem in a way that couldn't really be fixed without completely doing away with the egg benedict method.

=======

Analysis:

Looking at the issues with fantasy realm, part of the problem is that whomever wrote it assumed the function is100familiarRun() does what its name says instead of what it documentation says. this is not a break caused by my changes, but it was already broken before... however, to fix it i require a function that actually does what the name of is100familiarRun() says!
In addition, it also pointlessly repeated the mosquito check (which was already being checked in the function itself previously) resulting in incorrect data in paths like pokefam.

=======

In this commit:

used fine and replace alongside grep to completely replace all instances of 
is100FamiliarRun
with
autoForbidFamiliarChange

Now the naem is correctly self documenting. Moreoever, this frees up the name is100FamiliarRun for a function that actually correctly answers that question instead of answering the question of "am i forbidden to change familiar". Which I will now create, integrate, and go over all instances to see where the previous function was used incorrectly and is in need of switching to this new function.