created function is100FamiliarRun() which actually answers the question of the name. Modified autoForbidFamiliarChange() to make use of said function. Fixed documentation for both as well as autoForbidFamiliarChange(familiar thisOne) Also modified the structure of autoForbidFamiliarChange(familiar thisOne) to be clearer. instead of nested ifs use sequential ones to reach the same results but easier to read. now with extra documentation.
Created function autoUseFamiliar(familiar target), it is a wrapper to mafia's use_familiar(familiar target). it adds various checks to ensure proper functionality. it will return true if familiar has been successfully set to target, and false if it failed (either because it tried and failed, or because it realized you are forbidden from doing so).

Fixed fantasyRealmToken() in auto_mr2018.ash by
1. changing the check for "autoForbidFamiliarChange() + mosquito check" into a check for is100FamiliarRun().
Previously it would not run on any paths which forbid familiars except for pokefam. Now it will run on all paths that forbid familiars.
2. changing use_familiar($familiar[none]) into autoUseFamiliar($familiar[none]). which will prevent it from erroring out on pokefam path.

Also replaced 2 instances in autoscend.ash and 1 instance of auto_cooking.ash of use_familiar($familiar[none]) into autoUseFamiliar($familiar[none]). which should solve similar errors.