familiar functions changes (#317)
* familiar functions changes:
replaced forbidFamChange() with canChangeFamiliar()
replaced forbidFamChange(familiar target) with canChangeToFamiliar(familiar target)
This is not just a rename, but an inversion of the logic from asking a negative to asking a positive. 
Adjusted the functions to have inverted logic and all instances where those functions were called. (by adding or removing ! as appropriate)

*Fixed/improved various lines that called on these functions. With the refactoring of these functions bugs are easier to see at a glance