hounddog change instead of allowing hounddog whenever we did NOT providePlusNonCombat it will only onl allow it when we DID providePlusCombat
friars quest does not need hardcoding of +item familiar as it does not use that. let autoscend choose best familiar

removed some hardcoded switching to item familiar AFTER adventuring was done in a zone (in the assumption that whatever the next zone is, it needs it too).

add noncombat and combat familiar lists

*refactored ProvidePlusCombat/NonCombat to:
**Do not quit with return false if we cannot remove an unwanted effect
**Check if we are done after every item
**Reduce redundant code for checking if we are done via subfunction
**account for our equipment providing bonus/malus even if we do not change it
**not assume our muffler for sneaky pete.
**use dat file for combat and noncombat familiars. this means we now actually use disgeist for noncombat (hound dog was already used before for combat)
*buffmaintain will now verify that we got the right muffler before using muffled
*added support for unmuffled
*added unmuffled to providePlusCombat

improved code for familiar handling in provide plus combat/non from previous commit