fix infinite loop if you possess a hand in glove and try to acquireHP() (#529)
fix infinite loop if you possess a hand in glove and try to acquireHP()
owning a hand in glove breaks maxHP tracking when equipping, unequipping, or smithness score changes.
We need to check possession rather than equipped because unequipping it also breaks it. unequipping specifically can get autoscend stuck in an infinite loop of trying to restore hp when already at max HP because our actual maxHP is lower than we think it is.
mafia devs say that because the issue stems from a kol bug they won't add handling on their side
https://kolmafia.us/showthread.php?25214
so we need to work around this issue in autoscend.
The workaround is CLI command refresh status at the start of acquireHP function if we possess a hand in glove.