do not skip unlocking 2nd floor of spookyraven manor just because you have ghost of a necklace (#531)
do not skip unlocking 2nd floor of spookyraven manor just because you have ghost of a necklace.
this seems to be meant as error correction based on the logic that if we have a ghost of a necklace it means we must have finished this quest already. However we already use mafia tracking on this quest to check the state of it. And you can have ghost of a necklace either as a manual pull in softcore or as a result of being in casual.