fix quest softblock not releasing when level 13+ (#548)
the softblock release on quests that await optimal conditions is done after the return false if over level 12. this means that if you are level 13+ autoscend will never release the softblock. all those quests will continue waiting... thus it reaches the last line and quits saying it has nothing to do