if we decide to not do aboo peak then we should not equipBaseline() (#557)
*if we decide to not do aboo peak then we should not equipBaseline() but resetState() instead before returning false
the former actually changes our outfit which can cause us to lose HP from switching the outfit back and forth and waste a lot of meat. the latter does not do that and instead resets our maximizer as well as a few other things to as if we just started a new loop.