account for ed being undying when avoiding death by gremlin (#564)
account for ed being undying when avoiding death by gremlin.
#556 prevents us from committing suicide against gremlins via poking them with a seal tooth when they are projected to kill us within 1.5 rounds (a little over 1 round to account for inaccuracies in canSurvive())
However as ed the undying this is not an issue since he can resurrect for free twice per fight. will now account for this. will only do the above if not ed OR if ed is out of free resurrections