Avatar of Boris path fixes/support (#679)
* * boolean armorySubQuest() added
do the quest [Lending a Hand (and a Foot)] and unlock [madeline's baking supply] store

add setting to data auto_doArmory
for doing the army sidequest this ascension. by default enabled in boris and disabled elsewhere on initialize

* fix considerGalaktikSubQuest()
had the +100 on the wrong side of the equation. meat would never drop below reserve so it never fired.

* * boolean borisAcquireHP(int goal)
** boris cannot use the normal acquireHP function until it is modified allow multi using skills. that fix is nontrivial so until such a change is made here is a function that makes boris playable

* borisAcquireHP do not try to restore MP if maxMP is under 10

* * make sure we can equip an item before trying to wear it and visiting mourn