* ed catch island unlock failure. (#689)
ed catch island unlock failure. which is a critical issue with pathing ed that can easily turn a 3 day run into a 10 day run if not resolved in a timely manner.
failing to unlock the island should not happen. I have only seen it happen when it failed to initialize the path due to some mafia error. or because the user tried to play a few turns before running autoscend. in either case it will now catch this error and abort, informing the user on how to fix the problem before re-running autoscend.