kill ghosts if busting unavailable (#695)
kill ghosts if protonic pack is unavailable. which requires either:
1. a hypothetical future path that blocks back slot
2. using almost-dead walkie talkie to get a ghost without owning the IOTM.

walkie talkie is a very cheap item to get in the mall. as they are common drops from the proton pack and not that popular.
however almost-dead walkie talkie is not tracked by mafia correctly. when used the tracking does not update.
as such you need to perform a `refresh quests` command after using the walkie talkie.
autoscend automatically performs a single `refresh all` when it starts. so if a user used a walkie talkie and then ran autoscend then it will work