KGB fix infinite loop in bedtime if we can not find any buffs (#712)
fix infinite loop in kgbWasteClicks() when we are unable to find any valid buff to spend our clicks on
this usually happens if the user ran an external script to use the kgb.
this PR solves the infinite loop by detecting it and breaking the while function. It does not add special handling to spend your remaining clicks in such a scenario. That should be done in a future PR