Starting to work on free runs
At this point the auto_freeRunaway functions are just a barebones copy and paste of how voters and sausage goblins are handled and doesn't reflect how it will eventually have to look. 
The auto_banderRunsAvailable function is pretty solid in it's current state but I'm handling maximize calls in a bad way that doesn't agree with typical autoscend functionality and will need to be changed to work with eventual combat support (at the very least, survivability is necessary for %based free runs like tatters and if the same maximize call doesn't match then the number of runs could be off when combat is being run). 
The auto_freeRunsAvailable function is fine insofar as the logic is concerned but obviously there's way more that needs to be checked other than how many bander runs are available and tatters/GSBs. Also, I'm unsure if I want to include free banishes here since they're a resource that can be used elsewhere - if free runs are being burned at the end of the day, using free banishes is desirable BUT if free runs are being used before the end of the day to burn delay, using free banishes could be a bad thing.

I know how I want to handle bedtime logic, though I'm not sure where it will fall in the order of bedtime events. I don't know how to handle combat for free runs, but I'm working on figuring out how autoscend handles combat so I'll get there.