Add "Just the Best Anapests" to negative status effects
Added "Just the Best Anapests" to the negative status effects, as it appears to be shruggable and messes with mafia tracking. I had burned some spleen in QT to try to squeeze in some extra adventures on the last day of my run, but didn't shrug off the effect manually. I then burned 29turns in the palindome (disposable instant camera, used on turn 1 in that zone but not tracked) and 51 turns "Next to that Barrel with Something Burning in it" before getting the second abort and realizing something was up. 

Since this is shruggable, cosmetic flavor text that potentially interferes with tracking, can we shrug it off by default?