hotfix sauceror MP crash due to mortar delay (#844)
* hotfix sauceror MP crash (and subsequent meat crash as it has to keep on buying galaktik consumables to fuel its crashing MP) caused by a function that was trying to get more MP from the curse of weaksauce explosion. We were trying to let mortar deal the killing blow so we get extra MP.

to facilitate that we were delaying with salsaball, but accidentally killing the enemy with salsaball (especially at lower levels) and only getting back ~2MP. Thus losing lots of MP per combat instead of gaining MP. With the sauceror profligate spending of MP outside of combat (because of the assumption he has MP to spare) it combined into poverty.

Now it will prefer to delay with seal tooth with salsaball as a fallback. It will also check monster HP to make sure we do not kill it. Exact calculation is still being developed in #845 but this hotfix will just estimate things with making sure enemy HP is higher than 15 for salsaball delay. seal tooth will check >1 since it always deals 1 damage

* better tracking on which round we used mortar in to avoid delaying unnecessarily.