update outdated conditions for nightmare fuel (#846)
outdated condition was trying to use a spleen item if Powdered Gold is restricted. Now both that item and Powdered Gold are restricted together except in G lover. so autoscend produces the item but never uses it
I think it was trying to use it only in standard when there were no other uses for spleen. it makes more sense to use the same conditions to produce and use that item