* auto_bestWarPlan() fixed it was was only planning to do currently done sidequests + 1 undone sidequest. As sidequests were done more were added to the plan. This meant that it was not a plan at all but merely returning a value of "what is the nest best sidequest todo". only without considering that the sidequest in question might not be unlocked yet. particularly harmful for fighting for the hippies. The cause of the problem is weirdness in how records are handled. see below
* void copy_warplan(WarPlan target, WarPlan source) created and used
record A = B; does not copy the contents of B into record A, it instead copies memory references. Thus A merely becomes an alias for B and changing one changes the other as well.
this function serves to copy the data from B to into A
designed to be used with target.copy_warplan(source)