* hotfix auto_bestWarPlan() (#903)
* auto_bestWarPlan() fixed
it was was only planning to do the currently done sidequests + 1 undone sidequest with the best returns. After the best sidequest was done the next one was added to the plan. This meant that it was not a plan at all but merely returning a value of "what is the nest best sidequest todo". only without considering that the sidequest in question might not be unlocked yet. Which meant it was just wasting adventures by not doing a currently unlocked sidequest until it became the best value. The cause of this error is weirdness with how records are handled. see below

* void copy_warplan(WarPlan target, WarPlan source) created and used
record A = B; does not copy the contents of record B into record A, it instead copies memory address references from B into A. thus turning A into an alias of B and making it so that a change to one of them is automatically applied to the other going forwards.
this function serves to copy the data from B to into A while keeping them as separate variables.
designed to be used using the format `target.copy_warplan(source)`

* moved the block on hippies not doing arena (as this has not yet been implemented) to the consider stage. and clarified the comment explaining the block