int auto_mall_price(item it) fix (#909)
auto_mall_price has an issue if mall_price returns -1 for a tradeable item. this is usually caused by an IO error during mall search.
It results in an abort. in some situations (where we look at the price even for an item we already have) then autoscend will continue to abort the entire day and nothing you can do can make it run.

this PR will prevent the abort by using historical price instead, which is close enough in terms of value.
if this is the first time your mafia install tries to look up that item then historical price will return 0 even if it is available in the mall, which results in the item not being bought. potentially wasting some adventures. Although it is very much an edge case. As it requires a new mafia instance or your first time doing a new path and a path specific item to buy. As well as an IO error on that item's price lookup.