improve preparation for ghosts (#949)
improves pre adv preparation for ghosts. maximizing for prismatic damage was suboptimal since prismatic damage sources are rare and you are more likely to have specific elemental damage methods instead.
so now we will check the target location for physically immune monsters first. and then based on their element choose which elements to count and try to maximize for. using a simulation.