fix tomb already found issue (#990)
during nemesis quest the tomb of the unknown YOURCLASS only appears once when adventure in the unquiet garves.
we assumed that it has not been found yet and adventured in the unquiet garves for it. however if it was already exposed or if we found it but failed to complete it for some reason (ex: not having the correct item on hand) then it will continually waste adventures in the unquiet garves looking for the tomb even though it will not appear there anymore.
this PR solves this by checking to see if the tomb has been discovered already and if it has been going there directly.