Unique daily sc pulls (#1008)
Prevent multiple pulls of the same item in softcore each day. Part of support for nerf day changes 2022.

Can replace with whatever mafia's solution to this is later. Purpose of this change is to get initial support. No fun to stop part way through a day when consumable pulling fails

Fixes portion of #995