fix nightcap not working.
this was broken during #1293. in a function incorrectly calls itself instead of a different function. because it was not defined in header file.
mafia type checking does not apply to custom defined records. such that
```
test(string)
{return test(record)}
test(record)
```
would in fact have test(string) incorrectly calling itself. giving itself a record as if it was a string.

I added the missing functions to the header file autoscend/autoscend_record.ash
however, in order to make it possible to add said functions to move them to header file,
I needed to move the record in question to autoscend/autoscend_record.ash