Update auto_bedtime.ash
Not sure what the thoughts are toward the bedtime "resources are still available!" messages in this file; most are for stuff from back in 2015. But in that context, I've been running non-Standard runs and seeing reminders for Deck and Chateau use, I was considering the daily resources that are available that might be worth reminding users are available for manual intervention. 

A lot of these resources, even though they seem a bit wasted are also a bit trickier and maybe-not-advisable to tell a user to make use of (Carto mapping, making decisions about a random monster to reminisce about/fax, forcing NCs, clover uses), but I feel like August Scepter casts are a pretty fair one to remind users of?