feat: do not relocate lawyers to park in pokefam
It looks like behaviour was changed in #379 to always do this, instead
of only doing this if you were able to get Thrice-Cursed and were
forcing the elevator.

Regardless, this is a bad idea in general: noncombat forcers are
precious even when we have a lot of them, and if we're not forcing the
apartment takes 9 adventures to get to the office's 6, so reducing the
number of monsters by 1 in the office isn't worth it.

Pokefam cares a lot about the apartment because fighting shamans doesn't
give X-Cursed. Ideally, autoscend would get and drink Cursed Punch here.