fix: various pokefam improvements (#1676)
* fix: don't try to bullseye in pokefam

* fix: pokefam cannot free run either

* fix: cannot replace in pokefam either

* fix: pokefam starts with hippiesDefeated 500

* fix: pokefam blocks classic familiars

* fix: pokefam does not want disposable camera

* feat: do not relocate lawyers to park in pokefam

It looks like behaviour was changed in #379 to always do this, instead
of only doing this if you were able to get Thrice-Cursed and were
forcing the elevator.

Regardless, this is a bad idea in general: noncombat forcers are
precious even when we have a lot of them, and if we're not forcing the
apartment takes 9 adventures to get to the office's 6, so reducing the
number of monsters by 1 in the office isn't worth it.

Pokefam cares a lot about the apartment because fighting shamans doesn't
give X-Cursed. Ideally, autoscend would get and drink Cursed Punch here.

* fix: frat war in pokefam

* fix: cannot mctwist in pokefam

* fix: kill the wall of skin in pokefam

* fix: fight the shadow in pokefam

* fix: actually restore HP in pokefam, sometimes

* feat: configure leprecondo before adventuring in combat

* feat: buy pokefam hats

* feat: don't do shadow rifts at too low a level

* fix: only wish for trimmer to spend NC then fallbot if we have fallbot

* feat: wish for book of matches if we didn't get one

* fix: need to visit pokefam shop to get dollars

* fix: don't skip the beach as Frat

* fix: actually do the rifts, hopefully